+++ fp_analysis_app/__init__.py
@@ -5,5 +5,5 @@ Created on Mon Aug 12 11:50:57 2024
 @author: yzhao
 """
 
-VERSION = "v0.4.0"
+VERSION = "v0.4.1"
 DOWNSAMPLE_FACTOR = 100